Move hit testing & keyboard routing into gfx::input
