+++ crates/gfx/src/input/action.rs
@@ -1,0 +1,4 @@
+#[derive(Clone, Debug)]
+pub enum PageAction {
+    Navigate(String),
+}
